rename cli to cmd
